pyproject.toml
[project]
name = "ats-scanner"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0.0",
    "fastapi>=0.109.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-core>=0.1.10",
    "langchain-groq>=0.0.1",
    "pdfminer-six>=20221105",
    "python-docx>=0.8.11",
    "python-dotenv>=1.0.0",
    "python-multipart>=0.0.7",
    "sentence-transformers>=2.2.2",
    "spacy>=3.7.0",
    "uvicorn>=0.23.0",
]

[tool.uv.sources]
en-core-web-sm = { url = "https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl" }
